# Optimize row writer.