Summary:
Points of interest are basic_row_simple_tuple_assign,
and basic_row_simple_general.

The changes avoid the "need to commit" check for primitives.

Gap between vector and simple shrinks a lot (look at previous diff).
```
basic_row_vector                                            5.55ns   180.33M
basic_row_simple_tuple_assign                               6.15ns   162.70M
basic_row_simple_general                                    5.88ns   170.13M
array_of_rows                                             149.51ns     6.69M
complex_row                                               495.80ns     2.02M                                            502.26ns     1.99M
```

Differential Revision: D36350691